added jbuilder functions and set up the product object to be passed correctly

+++ db/migrate/20201115233139_create_products.rb
@@ -1,15 +1,15 @@
 class CreateProducts < ActiveRecord::Migration[6.0]
   def change
     create_table :products do |t|
-      t.int :sku
+      t.integer :sku
       t.string :name
-      t.double :price
-      t.double :weight
+      t.float :price
+      t.float :weight
       t.string :description
       t.string :thumbnail
       t.string :image
       t.string :category
-      t.int :stock
+      t.integer :stock
 
       t.timestamps
     end
