+++ db/migrate/20201115233626_create_order_details.rb
@@ -1,11 +1,11 @@
 class CreateOrderDetails < ActiveRecord::Migration[6.0]
   def change
     create_table :order_details do |t|
-      t.int :order_id
-      t.int :product_id
-      t.double :price
-      t.int :sku
-      t.int :quantity
+      t.integer :order_id
+      t.integer :product_id
+      t.float :price
+      t.integer :sku
+      t.integer :quantity
 
       t.timestamps
     end
